finished loadFile

+++ src/problemdomain/SUV.java
@@ -3,7 +3,7 @@ package problemdomain;
 public class SUV extends Vehicle{
 	
 
-	public SUV(int carID, String vehicleType, String subtype, int speed, float fuel, int seats, int year,
+	public SUV(long carID, String vehicleType, String subtype, int speed, double fuel, int seats, int year,
 			String driveTrain, int price, int quantity) {
 		super(carID, vehicleType, subtype, speed, fuel, seats, year, driveTrain, price, quantity);
 	}
